feat: upgrade to multi-sector RFID operations
- Increase key size from 6 bytes to 96 bytes (16 sectors x 6 bytes each)
- Increase payload size from 16 bytes to 512 bytes (16 sectors x 2 blocks x 16 bytes each)
- Update command parser to validate 192-char keys and 1024-char data
- Implement multi-sector read/write operations across sectors 0-15
- Read/write from blocks 1 and 2 of each sector (block 0 reserved for headers)
- Update bytesToHex function to handle larger data sizes (uint16_t length)
- Update documentation with new key and payload specifications
- Maintain power optimization and error handling for all operations

BREAKING CHANGE: Key and data sizes significantly increased - old commands incompatible

+++ include/RFIDController.h
@@ -24,7 +24,7 @@ private:
     bool powerUpNFC();
     void powerDownNFC();
     bool initializeNFC();
-    String bytesToHex(uint8_t *data, uint8_t length);
+    String bytesToHex(uint8_t *data, uint16_t length);
     void hexToBytes(const String &hex, uint8_t *bytes);
     bool authenticateBlock(uint8_t blockNumber, uint8_t *key);
 };
